+++ apps/react-frontend/src/features/roles/useAuth0Scopes.ts
